requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic~=2.11.5
playwright~=1.52.0
tortoise-orm[aiomysql]==0.25.1
python-dotenv